Move classes into folders
