--- Source/KiwiPizza/Private/LevelGenerator/MyLevel.cpp
@@ -1,12 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "LevelGenerator/MyLevel.h"
-
-MyLevel::MyLevel()
-{
-}
-
-MyLevel::~MyLevel()
-{
-}
