--- Source/KiwiPizza/Private/Pawns/MyPawn.cpp
@@ -1,12 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-
-#include "Pawns/MyPawn.h"
-
-MyPawn::MyPawn()
-{
-}
-
-MyPawn::~MyPawn()
-{
-}
