--- Source/KiwiPizza/Private/Pawns/MyPawn.h
@@ -1,15 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "CoreMinimal.h"
-
-/**
- * 
- */
-class MyPawn
-{
-public:
-	MyPawn();
-	~MyPawn();
-};
